requirements.txt
dnspython==2.6.1
PyQt6==6.6.1
PyQt6_sip==13.6.0
